Ajuste casas decimais na divisao

+++ primeiroprogramajavaeclipse/src/br/com/dio/calculadora/Calculadora.java
@@ -17,7 +17,7 @@ public class Calculadora {
 	int soma = soma(a,b);
 	int subtracao = subtracao(a,b);
 	int multiplicacao = mutiplicacao(a,b);
-	int divisao = divisao(a,b);
+	double divisao = divisao(a,b);
 	
 	System.out.println("Soma " + soma);
 	System.out.println("Subtracao " + subtracao);
@@ -39,7 +39,7 @@ public class Calculadora {
 		return  a * b;
 	}
 
-	public static int divisao(int a , int b) {
+	public static double divisao(double a , int b) {
 		return (a / b);
 	}
 }
